+++ src/Git_Updater/GU_Upgrade.php
@@ -30,7 +30,7 @@ class GU_Upgrade {
 	 *
 	 * @var int
 	 */
-	private $db_version = 8312;
+	private $db_version = 8312; // TODO: change number.
 
 	/**
 	 * Run update check against db_version.
@@ -38,6 +38,7 @@ class GU_Upgrade {
 	public function run() {
 		$options    = $this->get_class_vars( 'Base', 'options' );
 		$db_version = isset( $options['db_version'] ) ? (int) $options['db_version'] : 6000;
+		$this->schedule_access_token_cleanup();
 
 		if ( $db_version === $this->db_version ) {
 			return;
@@ -51,7 +52,13 @@ class GU_Upgrade {
 				break;
 		}
 
-		$options = array_merge( (array) $options, [ 'db_version' => (int) $this->db_version ] );
+		$options = array_merge(
+			(array) $options,
+			[
+				'db_version'     => (int) $this->db_version,
+				'gu_auto_update' => '1',
+			]
+		);
 		update_site_option( 'git_updater', $options );
 	}
 
@@ -76,7 +83,73 @@ class GU_Upgrade {
 			update_site_option( 'git_updater', $ghu_options );
 			delete_site_option( 'github_updater' );
 		}
-		require_once ABSPATH . '/wp-admin/includes/plugin.php';
+		require_once ABSPATH . 'wp-admin/includes/plugin.php';
 		\deactivate_plugins( 'github-updater/git-updater.php' );
+	}
+
+	/**
+	 * Check for deletion of cron event.
+	 *
+	 * @param null|bool|\WP_Error $pre       Value to return instead. Default null to continue unscheduling the event.
+	 * @param int                 $timestamp Timestamp for when to run the event.
+	 * @param string              $hook      Action hook, the execution of which will be unscheduled.
+	 *
+	 * @return null|bool|\WP_Error
+	 */
+	public function pre_unschedule_event( $pre, $timestamp, $hook ) {
+		if ( 'gu_delete_access_tokens' === $hook ) {
+			$days = ( \wp_next_scheduled( 'gu_delete_access_tokens' ) - time() ) / \DAY_IN_SECONDS;
+			if ( $days > 29 ) {
+				$this->flush_tokens();
+			}
+		}
+		return $pre;
+	}
+
+	/**
+	 * Update for non-password options.
+	 *
+	 * @since 12.0.0
+	 *
+	 * @return void
+	 */
+	public function flush_tokens() {
+		if ( gu_fs()->can_use_premium_code() || false === wp_next_scheduled( 'gu_delete_access_tokens' ) ) {
+			return;
+		}
+
+		$base_options = [
+			'db_version',
+			'branch_switch',
+			'bypass_background_processing',
+			'deprecated_error_logging',
+			'gu_auto_update',
+		];
+		$options      = $this->get_class_vars( 'Base', 'options' );
+		$new_options  = \array_filter(
+			$options,
+			function( $value, $key ) use ( &$options, $base_options ) {
+				if ( in_array( $key, $base_options, true ) || str_contains( $key, 'current_branch' ) ) {
+					return $options[ $key ];
+				}
+			},
+			ARRAY_FILTER_USE_BOTH
+		);
+		update_site_option( 'git_updater', $new_options );
+	}
+
+	/**
+	 * Schedule cleanup of the access tokens.
+	 *
+	 * @since 12.0.0
+	 *
+	 * @return void
+	 */
+	private function schedule_access_token_cleanup() {
+		if ( false === wp_next_scheduled( 'gu_delete_access_tokens' ) ) {
+			wp_schedule_event( time() + \MONTH_IN_SECONDS, 'twicedaily', 'gu_delete_access_tokens' );
+		}
+
+		add_action( 'gu_delete_access_tokens', [ $this, 'flush_tokens' ] );
 	}
 }